rename shader string preprocessing file
